pyproject.toml
[project]
name = "vogel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "django-environ>=0.12.0",
    "django-ninja>=1.3.0",
    "psycopg2>=2.9.10",
]
